% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gct_ppt.R
\name{make_gct_ppt}
\alias{make_gct_ppt}
\title{Populate Grad Club Trivia Powerpoint Function}
\usage{
make_gct_ppt(questions, template, name = "Thea")
}
\arguments{
\item{questions}{An xlsx spreadsheet with at least two columns: Question and QuestionN}

\item{template}{The powerpoint template. Defaults to GCTrivia template (includes cover slide, font, and color scheme)}
}
\description{
This function takes an ordered spreadsheet of trivia questions, the GC powerpoint template, and populates a presentation with questions.
}
