% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gct_slide.R
\name{make_gct_slide}
\alias{make_gct_slide}
\title{Make Grad Club Trivia slide}
\usage{
make_gct_slide(q, qn, pres)
}
\arguments{
\item{q}{Vector containing questions}

\item{qn}{Vector containing question numbers}

\item{pres}{Powerpoint object on which to add slides}
}
\description{
Make Grad Club Trivia slide
}
