% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ppt.R
\name{make_ppt}
\alias{make_ppt}
\title{Populate Powerpoint Function}
\usage{
make_ppt(questions, purpose = "gct")
}
\arguments{
\item{questions}{An xlsx spreadsheet with at least two columns: Question and QuestionN OR a txt file with the first row labelled Question (TODO: fix)}

\item{type}{Type of project. Defaults to lc (LOUD Crowd) template.}
}
\description{
This function takes an ordered spreadsheet of trivia questions (gct) OR a list of sentences (lc) and the type of project (gct or lc) and populates a presentation.
}
