% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_slide.R
\name{make_slide}
\alias{make_slide}
\title{Make Grad Club Trivia slide}
\usage{
make_slide(q, qn, pres, m)
}
\arguments{
\item{q}{Vector containing questions}

\item{qn}{Vector containing question numbers}

\item{pres}{Powerpoint object on which to add slides}

\item{m}{Master name}
}
\description{
Make Grad Club Trivia slide
}
